New Extension attributes
Holy cow, a commit! It's my first one in over a year and it's a boring one..and boy, I am rusty!

New EAs created to scan the Users folder to find VM preference files and try to extrapolate the guest OS based on some settings.

It's ugly, but it works for what I need right now.